% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_debug.R
\name{gn_debug}
\alias{gn_debug}
\title{gn_parse}
\usage{
gn_debug(x, threads = 4)
}
\arguments{
\item{x}{(character) vector of scientific names}

\item{threads}{(integer/numeric) xxx. default: 4}

\item{format}{(logical) xxx. default: "compact"}
}
\value{
list of two data.frames:
\itemize{
\item meta: metadata
\item names: names and their parts, varies based on function parameters
}
}
\description{
extract names using gnparser
}
\examples{
z <- c("Quadrella steyermarkii (Standl.) Iltis &amp; Cornejo",
  "Parus major Linnaeus, 1788", "Helianthus annuus var. texanus")
x <- gn_debug(z[1])
x
gn_debug(z[2])
gn_debug(z[3])
}
