% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_parse.R
\name{gn_parse}
\alias{gn_parse}
\title{gn_parse}
\usage{
gn_parse(x, format = "compact", threads = 4)
}
\arguments{
\item{x}{(character) vector of scientific names}

\item{format}{(logical) xxx. default: "compact"}

\item{threads}{(integer/numeric) xxx. default: 4}
}
\value{
list of two data.frames:
\itemize{
\item meta: metadata
\item names: names and their parts, varies based on function parameters
}
}
\description{
extract names using gnparser
}
\examples{
x <- c("Quadrella steyermarkii (Standl.) Iltis &amp; Cornejo",
  "Parus major Linnaeus, 1788", "Helianthus annuus var. texanus")
gn_parse(x[1])
gn_parse(x[2])
gn_parse(x[3])
}
