% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_version.R
\name{gn_version}
\alias{gn_version}
\title{gn_version}
\usage{
gn_version()
}
\value{
prints version information to console, nothing returned
}
\description{
get gnparser version information
}
\examples{
gn_version()
}
