% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gnparser.R
\name{install_gnparser}
\alias{install_gnparser}
\title{Install gnparser}
\usage{
install_gnparser(version, force)
}
\arguments{
\item{version}{The gnparser version number, e.g., \verb{1.0.0}; the default
\code{latest} means the latest version (fetched from GitLab releases).
Alternatively, this argument can take a file path of the zip archive or
tarball of gnparser that has already been downloaded from GitLab,
in which case it will not be downloaded again. The minimum version
is \code{v1.0.0} because gnparser v1 introduced breaking changes - and
we don't support older versions of gnparser here.}

\item{force}{Whether to install gnparser even if it has already been
installed. This may be useful when upgrading gnparser.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\subsection{Reason for deprecating}{

The function used to download the appropriate \code{gnparser} executable for your
platform and try to copy it to a system directory so \pkg{rgnparser} can run
the \code{gnparser} command.
This function was deprecated to stay in compliance with CRAN policies
("Packages should not write in the user’s home filespace, nor anywhere else
on the file system apart from the R session’s temporary directory")
}

\subsection{Solution}{

Please install \code{gnparser} by hand.
For Linux and Mac users, installing with your usual package manager such as
homebrew is the easiest, see \code{gnparser} documentation for more details:
\url{https://github.com/gnames/gnparser#installation}
}
}
