% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_debug.R
\name{gn_debug}
\alias{gn_debug}
\title{gn_debug}
\usage{
gn_debug(x, threads = 4)
}
\arguments{
\item{x}{(character) vector of scientific names}

\item{threads}{(integer/numeric) number of threads. default: 4}
}
\value{
cats output to screen
}
\description{
debug parsing
}
\examples{
z <- c("Quadrella steyermarkii (Standl.) Iltis &amp; Cornejo",
  "Parus major Linnaeus, 1788", "Helianthus annuus var. texanus")
x <- gn_debug(z[1])
x
unclass(x)
}
