% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_parse_tidy.R
\name{gn_parse_tidy}
\alias{gn_parse_tidy}
\title{gn_parse_tidy}
\usage{
gn_parse_tidy(x, threads = 4)
}
\arguments{
\item{x}{(character) vector of scientific names}

\item{threads}{(integer/numeric) number of threads to run. CPU's
threads number is the default. default: 4}
}
\value{
a data.frame
}
\description{
extract names using gnparser into a tidy tibble
}
\examples{
trys <- function(x) try(x, silent=TRUE)
if (interactive()) {
x <- c("Quadrella steyermarkii (Standl.) Iltis &amp; Cornejo",
  "Parus major Linnaeus, 1788", "Helianthus annuus var. texanus")
trys(gn_parse_tidy(x))
}
}
