% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_version.R
\name{gn_version}
\alias{gn_version}
\title{gn_version}
\usage{
gn_version()
}
\value{
named list, with \code{version} and \code{build}
}
\description{
get gnparser version information
}
\examples{
trys <- function(x) try(x, silent=TRUE)
if (interactive()) {
trys(gn_version())
}
}
